ai_researcher/pyproject.toml
[project]
name = "ai-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chroma",
    "chromadb>=1.2.1",
    "db",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.119.1",
    "fastmcp>=2.12.5",
    "langchain",
    "langchain-community",
    "langchain-google-genai>=3.0.1",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.22.0",
    "typing-extensions>=4.15.0",
    "wheel>=0.45.1",
]

[tool.uv.workspace]
members = [
    "app/core",
]


